requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple/
amqp==5.0.9; python_version >= '3.6'
apscheduler==3.6.3
backports.zoneinfo==0.2.1; python_version >= '3.6' and python_version < '3.9'
billiard==3.6.4.0
cachetools==4.2.2; python_version ~= '3.5'
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11; python_version >= '3'
click-didyoumean==0.3.0; python_version < '4' and python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.2.0
click==8.0.3; python_version >= '3.6'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3'
kombu==5.2.3; python_version >= '3.7'
packaging==21.3; python_version >= '3.6'
prompt-toolkit==3.0.26; python_full_version >= '3.6.2'
pycparser==2.21
pygithub==1.55
pyjwt==2.3.0; python_version >= '3.6'
pynacl==1.5.0; python_version >= '3.6'
pyparsing==3.0.7; python_version >= '3.6'
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytz==2021.3
redis==4.1.2
requests==2.27.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tornado==6.1; python_version >= '3.5'
tzdata==2021.5; python_version >= '3.6'
tzlocal==4.1; python_version >= '3.6'
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
wrapt==1.13.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
